+++ collab/src/index.ts
@@ -3,7 +3,7 @@ import type { Request, Response } from "express";
 /* eslint "import/no-named-as-default-member": "off" -- This is the only way I can see to avoid a warning about the (correct) way express is imported below */
 import express from 'express';
 import {authMiddleware} from "./middleware/auth-middleware";
-import { validateSteps } from "./lib/parse-steps";
+import { validateSteps } from "./lib/validate-steps";
 
 export const app = express();
